Backup and restore utilities for Roboquest configuration files (#109)
* Added fastcgi support of Python scripts to the registry lighttpd
configuration.
Added a fastcgi Python script which can read an incoming POST
request and extract some meta data.

https://github.com/billmania/roboquest_core/issues/108

* POST a file and repost the same file to the backup directory.
Incorrectly returns a status 200 on error.

https://github.com/billmania/roboquest_core/issues/108

* Secured the file name in python-fcgi for the POST payload.
Added backup.sh to POST the config files.

https://github.com/billmania/roboquest_core/issues/108

* Added restore.sh, to retrieve the archive and restore its contents.
Changed the location of the archives on the registry, from backups
to archives.
Changed the content-type in backup.sh to application/x-gtar-compressed
and ignore missing files.

https://github.com/billmania/roboquest_core/issues/108

* Fixed a bug with the naming of the retrieved tar archive in
restore.sh.

https://github.com/billmania/roboquest_core/issues/108